Fix failing tests
The previous implementation assumed that only one lockRenewer
timer is active at a time, but there can be several. The timer
manager has been updated to reflect this.

In addition, the delay timer is no longer managed by the timer
manager since it is persistent (its initial value is
Number.MAX_VALUE) and would never be cancelled. Instead,
it is closed immediately by close().