Add support for rate limiting
Fix #362
Fix #329

This commit adds support for rate limiting of a queue. The rate limit
applies to each instance of the queue. If you have 3 threads processing
the queue, and you want to limit your jobs to 1 per second, it will be
1 job per second amongst the 3 threads.

queue#process first argument has been changed to an options object, and
concurrency now resides within this object, as well as the new rate
limiting options. (If a number is specified as the first argument to
queue#process, it will be treated as the old version of the API which I
have deprecated in favor of passing in the options object).