Consider jobs being retrieved when pausing queue locally.
It was possible for our queue to be "paused", but still be working
on jobs, because we were not considering jobs that were in the
process of being retrieved when the pause was initiated.

This ensures that our queue is not considered paused until we have
stopped listening for jobs to process, and until we have processed
all jobs that we might have retrieved after initiating the pause.