Only wait for one job at a time.
When waiting for a new job, we generally use a blocking redis
command, which means that no other redis commands will be executed
until the blocking command has completed.

If we try to execute any commands while the blcoking command is
active, ioredis will queue those commands and execute them after
the blocking command completes.

When we do `queue.process(n, function...`, we issue `n` blocking
commands to wait for new jobs, one for each job we wish to process.
But, since this is a blocking command, only the first is actualy
executed, and the rest are queued up to be executed later.

When we want to pause our queue, we need to wait for these blocking
commands to complete, otherwise we might end up processing a job
when we think the queue is paused.

If we queue all of the blocking commands up front, then it can take
a long time for them to all complete one at a time.

This changes the code to only issue one blocking command at a time,
and if there is already one pending, wait until it resolves before
issuing another blocking command.

This means that we can re-check the state of our queue more often,
and if we find that our queue is pausing, we can choose not to
issue the blocking command to wait for another job to process.