fix(queue): fix calculation for next repeat job
Fix `delay` calculation. Previously, it would not work if `queue.add`
was called multiple times, since the delay would always be the
interval length instead of the time from now

Add `isRepeat` flag to change which iteration is calculated. Only
calculate based on queue `millis` when we repeat to get the next job,
otherwise use the current time so the immediate next interval is used