docs: added @breejs
I created @breejs as a simpler alternative to Bull.  Just in case someone doesn't need something as advanced as Bull, it might be helpful to share.

Bull can even be used inside of Bree itself, in one of the workers.  For example, you could have a Bree worker that spawns a Bull queue!