chore: ioredis 4.18.0 → ^4.19.2
Update from pinned ioredis 4.18.0 to the latest version via ^4.19.2.

ioredis was previously pinned to 4.18.0 after tests began breaking with
the upgrade to 4.19.0.

Running tests with this version should result in a number of failures
in the repeatable job test suite.

These failures are a result of the manner in which `sinon` is used to
make massive advancements in time, thus triggering huge amounts of
interval handler executions as catch up is done.