fix(queue): fix canceling job with timeout or non-native Promise
We were wrapping the job promise a couple times, and thus clobbering any
cancel() method that was provided. This wraps more carefully, to fix it
for at least in-process jobs.

It did work previously if you added a `cancel` method to a native
promise for a job without a timeout, because `Promise.resolve()` just
returns the promise itself if it is a Promise object.

At least partially fixes #2203, this fix only focuses on in-process
jobs, and does not affect separate-process jobs.