Update REFERENCE.md to note a performance footgun
At my company, we ran `queue.clean(0, "waiting");` on a queue with 1.5 million tasks enqueued (don't ask, cleaning up other peoples' messes is my specialty). This worked locally almost instantly on a couple thousand tasks. But in our lower environments, the Lua script it runs pinned our Redis server to 100% CPU for **over 18 hours** while it worked. The database stopped responding to any commands, and we could not shut it off, even through the GCP hosted databases console.

We had to replace our Redis database (the service stopped responding entirely) for the environment. Luckily, we did not run this on production, or we would have been EXTRA boned. I only blocked the company for around 3 hours before we were able to recover. 😢